+++ packages/adminui-vue/components/UbTabbar/init.js
@@ -1,8 +1,9 @@
 const Vue = require('vue')
 const UbTabbar = require('./src/index.vue').default
+const UB = require('@unitybase/ub-pub')
 
 module.exports = function () {
-  const { id } = window.UB.core.UBApp.viewport.centralPanel.tabBar
+  const { id } = UB.core.UBApp.viewport.centralPanel.tabBar
   const styles = document.createElement('style')
   styles.innerHTML = `#${id}{display:none !important}`
   document.body.appendChild(styles)
@@ -13,6 +14,10 @@ module.exports = function () {
       window.UB.core.UBApp.viewport.centralPanel.setMargin(`-${offsetHeight} 0 0 0`)
       window.UB.core.UBApp.viewport.centralPanel.tabBar.setHeight(offsetHeight)
     },
-    render: (h) => h(UbTabbar)
+    render: (h) => h(UbTabbar, {
+      props: {
+        withHamburger: UB.connection.appConfig.uiSettings.adminUI.customSidebar
+      }
+    })
   }).$mount(`#${id}`)
 }